Partly fix art with sabermod 4.8
First error fixed - now there's another error:
F/art     (  604): art/runtime/well_known_classes.cc:118] Couldn't find method "nativeLoad" with signature "(Ljava/lang/String;Ljava/lang/ClassLoader;Ljava/lang/String;)Ljava/lang/String;"
But I have no idea how to fix that. I never used JNI and therefore don't know how it works.
The signature seems to be right as can be seen in runtime/native/java_lang_Runtime.cc => search for "NATIVE_METHOD".

+++ runtime/mirror/art_method.h
@@ -155,7 +155,7 @@ class MANAGED ArtMethod : public Object {
   }
 
   // Number of 32bit registers that would be required to hold all the arguments
-  static size_t NumArgRegisters(const StringPiece& shorty);
+  static uint32_t NumArgRegisters(const StringPiece& shorty);
 
   uint32_t GetDexMethodIndex() const;
